# PERF: 5x speedup for read_json() with orient='index' by avoiding transpose