--- a/pandas/io/json/_json.py
+++ b/pandas/io/json/_json.py
@@ -1085,9 +1085,15 @@ class FrameParser(Parser):
             self.check_keys_split(decoded)
             self.obj = DataFrame(dtype=None, **decoded)
         elif orient == "index":
-            self.obj = DataFrame(
-                loads(json, precise_float=self.precise_float), dtype=None
-            ).T
+            self.obj = (
+                DataFrame.from_dict(
+                    loads(json, precise_float=self.precise_float),
+                    dtype=None,
+                    orient="index",
+                )
+                .sort_index(axis="columns")
+                .sort_index(axis="index")
+            )
         elif orient == "table":
             self.obj = parse_table_schema(json, precise_float=self.precise_float)
         else: