The `.T` operator can be quite slow on mixed-type `DataFrame`s due to the creation of `object` dtype columns. In comparison to direct construction with `DataFrame.from_dict()` can generally be much more efficient.

Making that swap inside `pd.read_json()` yields a `~5-6x` speedup for the `orient='index'` case: 
```
       before           after         ratio
     [d47fc0cb]       [b0fd99ec]
     <read_json_speedup~1>       <read_json_speedup>
-      5.37±0.03s          907±5ms     0.17  io.json.ReadJSON.time_read_json('index', 'int')
-      5.27±0.01s          804±3ms     0.15  io.json.ReadJSON.time_read_json('index', 'datetime')
```

- [ ] closes #xxxx
- [ ] tests added / passed
- [ ] passes `git diff upstream/master -u -- "*.py" | flake8 --diff`
- [ ] whatsnew entry