@WillAyd 2019-06-10
cc @TomAugspurger would this be related to https://github.com/pandas-dev/pandas/pull/24387 at all?