@WillAyd 2019-06-10
Ignore previous comment was too focused on the constructor and not the transition. This makes sense to me