@qwhelan 2019-06-10
@alimcmaster1 Given that this only fails on 3.5, I'm guessing this is a dict-orderedness issue in `from_dict()`