@jreback 2019-07-08
lgtm, can you add a note in Performance for 0.25.0, ping on green.